Added 'where' conditions to 'delete from' statement (#1624). Documented.
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@4307 406a0c4d-033a-0410-8de8-e80135713968

+++ src/llsphinxql.c
@@ -1004,7 +1004,7 @@ static int input (yyscan_t yyscanner );
 	if ( YY_CURRENT_BUFFER_LVALUE->yy_is_interactive ) \
 		{ \
 		int c = '*'; \
-		int n; \
+		size_t n; \
 		for ( n = 0; n < max_size && \
 			     (c = getc( yyin )) != EOF && c != '\n'; ++n ) \
 			buf[n] = (char) c; \
